fix: make changes based on feedback
