+++ packages/base-driver/test/basedriver/helpers-specs.js
@@ -1,7 +1,6 @@
-import { isPackageOrBundle, duplicateKeys, parseCapsArray, parseServerArgs } from '../../lib/basedriver/helpers';
+import { isPackageOrBundle, duplicateKeys, parseCapsArray } from '../../lib/basedriver/helpers';
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import _ from 'lodash';
 
 
 chai.use(chaiAsPromised);
@@ -108,48 +107,5 @@ describe('parseCapsArray', function () {
   });
   it('should return an array without change', function () {
     parseCapsArray(['a', 'b']).should.eql(['a', 'b']);
-  });
-});
-
-describe('parseServerArgs', function () {
-  const webkitDebugProxyPort = 22222;
-  const wdaLocalPort = 8000;
-  const driverArgs = { wdaLocalPort, webkitDebugProxyPort };
-  const opts = {'foo': 'bar', 'foobar': 'foobar'};
-  const ARGS_CONSTRAINTS = {
-    webkitDebugProxyPort: {
-      isNumber: true
-    },
-    wdaLocalPort: {
-      isNumber: true
-    },
-  };
-
-  it('should return driver args if passed in', function () {
-    parseServerArgs({}, driverArgs, ['wdaLocalPort', 'webkitDebugProxyPort'], ARGS_CONSTRAINTS).should.eql(driverArgs);
-  });
-  it('should assign driver args to opts if passed in', function () {
-    parseServerArgs(opts, driverArgs, ['wdaLocalPort', 'webkitDebugProxyPort'], ARGS_CONSTRAINTS).should.eql(_.assign(opts, driverArgs));
-  });
-  it('should use opts args if driver args not passed in', function () {
-    parseServerArgs(_.assign(opts, driverArgs), {}, ['wdaLocalPort',
-      'webkitDebugProxyPort'], ARGS_CONSTRAINTS).should.eql(_.assign(opts, driverArgs));
-  });
-  it('should return empty object if no args were passed in', function () {
-    parseServerArgs({}, {}, ['wdaLocalPort', 'webkitDebugProxyPort'], ARGS_CONSTRAINTS).should.eql({});
-  });
-  describe('wdaLocalPort arg', function () {
-    it('should return empty obj if driverArgs is empty', function () {
-      parseServerArgs({}, {}, ['wdaLocalPort'], ARGS_CONSTRAINTS).should.eql({});
-    });
-    it(`should throw error if value of 'wdaLocalPort' is not an int`, function () {
-      (() => parseServerArgs({}, {'wdaLocalPort': 'foo'}, ['wdaLocalPort'], ARGS_CONSTRAINTS)).should.throw();
-    });
-    it(`should throw if unreconized key is passed`, function () {
-      (() => parseServerArgs({}, {'foo': 'bar'}, ['wdaLocalPort'], ARGS_CONSTRAINTS)).should.throw();
-    });
-    it('should return passed in driver arg value', function () {
-      parseServerArgs({}, {wdaLocalPort}, ['wdaLocalPort'], ARGS_CONSTRAINTS).wdaLocalPort.should.equal(wdaLocalPort);
-    });
   });
 });
